% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_linestring_from_json.R
\name{extract_linestring_from_json_list}
\alias{extract_linestring_from_json_list}
\title{Extract linestring geometry from a list with a json-like structure}
\usage{
extract_linestring_from_json_list(feature_iterator, json_list)
}
\arguments{
\item{feature_iterator}{An integer indexing the feature element to extract}

\item{json_list}{A list with json structure containing linestring coordinates}
}
\value{
An sf linestring object
}
\description{
Extract linestring geometry from a list with a json-like structure
}
\examples{

json_list <- list(
features=list(
list(
 geometry = list(
   coordinates =list(
     c(0,0),c(0,1),c(1,1),c(1,0),c(0,0)
   )
 )
),
  list(
 geometry = list(
   coordinates =list(
     c(0,0),c(0,3),c(3,0),c(0,0)
   )
 )
),
list(
 geometry = list(
   coordinates =list(
     c(-1,0),c(1,0),c(1,1),c(0,2),c(-1,1),c(-1,0)
   )
 )
)
)
 )

 extract_linestring_from_json_list(3,json_list)
}
