% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_unpivot.R
\name{parallel_unpivot}
\alias{parallel_unpivot}
\title{Simlutanously unpivot two variables running "parallel" across a wide-format data frame columns}
\usage{
parallel_unpivot(
  raw_df,
  start_pattern1,
  start_pattern2,
  new_name1,
  new_name2,
  index_name,
  regex_group
)
}
\arguments{
\item{raw_df}{A dataframe in \strong{wide format}}

\item{start_pattern1}{\strong{Character} identifying first variable}

\item{start_pattern2}{\strong{Character} identifying second variable}

\item{new_name1}{New name for the column storing the values of the first variable}

\item{new_name2}{New name for the column storing the values of the second variable}

\item{index_name}{New name for the column storing the \strong{index} (usually time, e.g. hours, minutes) that \strong{varies simultaneously} through both variables across the input dataframe's columns}

\item{regex_group}{Regular expression containing \strong{matching groups} using parentheses, to split the input dataframe's columns into two parts: the identifying character for each of the two variable and the varying index.}
}
\value{
A data frame in long format with three new columns: one for the varying index and two for each of the "parallel" variables
}
\description{
Simlutanously unpivot two variables running "parallel" across a wide-format data frame columns
}
\examples{
df <- data.frame(a=1:5,
b=6:10,
X1=rep(1,5),
Y1=rep(10,5),
X2=rep(2,5),
Y2=rep(20,5),
X30=rep(3,5),
Y30=rep(30,5)
)

parallel_unpivot(df,'X','Y','X_series','Y_series','index','(X|Y)([[:digit:]]+)')
}
