% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename the column names of a dataframe without calling any external library}
\usage{
rename_columns(df, old_colnames, new_colnames)
}
\arguments{
\item{df}{A dataframe}

\item{old_colnames}{A \strong{character vector} with the column names  of the data frame \strong{to be changed}.}

\item{new_colnames}{A \strong{character vector} with the new column names. Each element should match the corresponding \strong{old name} in the \code{old_names} parameter.}
}
\value{
A dataframe of the same dimensions, with changed column names.
}
\description{
Rename the column names of a dataframe without calling any external library
}
\examples{
df <- data.frame(a=rep(1,5),b=rep(2,5),c=rep(3,5))
rename_columns(df,'a','first_column')
#
#
#' rename_columns(df,c('b','c'),c('second_column','third_column'))
}
